+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-*
+* It's now possible to configure the size modal's select button label both through props and `open_size` event options
 
 ### Changed