+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * As default, extracts the reference size and available scales from the 3DB's `spec.json`
 * It's now possible to configure the size modal's select button label both through props and `open_size` event options
+* Configurable labels for personalization/size buttons and modals through the store
 
 ### Changed